- Other formats should not cut off now. - mem_fsize should return size_t
SVN r730 (trunk)

+++ common/m_memio.cpp
@@ -210,7 +210,7 @@ int mem_fseek(MEMFILE *stream, signed long position, mem_rel_t whence)
 	}
 }
 
-long mem_fsize(MEMFILE *stream) // [Russell] - get size of stream
+size_t mem_fsize(MEMFILE *stream) // [Russell] - get size of stream
 {
     return stream->buflen;
 }
