+++ common/m_memio.h
@@ -43,7 +43,7 @@ void mem_get_buf(MEMFILE *stream, void **buf, size_t *buflen);
 void mem_fclose(MEMFILE *stream);
 long mem_ftell(MEMFILE *stream);
 int mem_fseek(MEMFILE *stream, signed long offset, mem_rel_t whence);
-long mem_fsize(MEMFILE *stream); // [Russell] - get size of stream
+size_t mem_fsize(MEMFILE *stream); // [Russell] - get size of stream
 char *mem_fgetbuf(MEMFILE *stream); // [Russell] - return stream buffer
 #endif /* #ifndef MEMIO_H */
